Move types to types.d.ts
