+++ src/PlayerLoadErrorMessage.tsx
@@ -1,6 +1,5 @@
 import React, { CSSProperties } from 'react'
-
-import { LoadErrorMessageProps } from './types'
+import { LoadErrorMessageProps } from '../types'
 
 export default function ErrorMessage(props: LoadErrorMessageProps) {
   const containerStyle: CSSProperties = {
